requirements.txt
# fastapi==0.115.8
# uvicorn==0.24.0.post1
# databases[postgresql]==0.8.0
# asyncpg==0.30.0
# SQLAlchemy==1.4.50
# python-dotenv==1.0.0
# passlib[bcrypt]==1.7.4
# pyjwt==2.7.0
# pydantic==2.10.6
# pydantic[email]
# python-jose
# python-multipart
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
# passlib[bcrypt]
python-dotenv
python-jose
pydantic[email]
python-multipart
email-validator
httpx
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
starlette
odoorpc
requests
boto3
stripe
alembic
psycopg2-binary
boto3
chromadb
PyPDF2
python-docx
openpyxl
tqdm
openai>=1.0.0
pillow
pytesseract
opencv-python-headless
numpy